Move ifixit product api call to its own file.
