+++ frontend/lib/ifixit-api/productData.ts
@@ -1,0 +1,28 @@
+import { IFixitAPIClient } from '@ifixit/ifixit-api-client';
+import { invariant } from '@ifixit/helpers';
+
+export type ProductDataApiResponse = {
+   variantOptions: {
+      [o_code: string]: string[];
+   };
+};
+
+export type VariantOptions = {
+   [o_code: string]: string[];
+};
+
+export async function fetchProductData(
+   client: IFixitAPIClient,
+   handle: string
+): Promise<ProductDataApiResponse | null> {
+   const productHandle = encodeURIComponent(handle);
+   try {
+      invariant(
+         productHandle.length > 0,
+         'productHandle cannot be a blank string'
+      );
+      return await client.get(`store/product/${productHandle}`);
+   } catch (error: any) {
+      return null;
+   }
+}
